ASoC: wcd-mbhc: Increase number of measurements for impedance calc to 900
Stereo headset sometimes gets detected as mono headset.
This is due to incomplete ramp on headphone left and/or right.
Increase the retry loop counter to 900 so as to give
more time for the ramp to be completed.

CRs-Fixed: 2084750
Change-Id: Ibe33e4e8fe08405a8b30bf427742943b91491ef8
Signed-off-by: Vatsal Bucha <vbucha@codeaurora.org>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>
(cherry picked from commit d7ed94be30aed45e1bb48b40c4505ca71436dcae)